requirements.txt
django
celery
redis
langchain
langchain-community
langchain-openai
openrouter
djangorestframework
drf-spectacular
django-celery-results
flower
pytz
gunicorn
uvicorn[standard]
psycopg2-binary
